Add support for Symfony 7
