+++ .github/workflows/license-audit.yml
@@ -7,7 +7,11 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
+        php-version: ['8.0']
         symfony-version: ['^3.4', '^4.4', '^5.0', '^6.0']
+        include:
+            - symfony-version: '^7.0'
+              php-version: '8.3'
 
     steps:
     - uses: actions/checkout@v2
@@ -15,7 +19,7 @@ jobs:
     - name: install PHP
       uses: shivammathur/setup-php@v2
       with:
-        php-version: '8.0'
+        php-version: ${{ matrix.php-version }}
 
     - name: Fetch decisions.yml
       run: curl https://raw.githubusercontent.com/bugsnag/license-audit/master/config/decision_files/global.yml -o decisions.yml
