+++ .github/workflows/test-package.yml
@@ -86,12 +86,16 @@ jobs:
             symfony-version: '^5.0'
           - php-version: '8.2'
             symfony-version: '^6.0'
+          - php-version: '8.2'
+            symfony-version: '^7.0'
           - php-version: '8.3'
             symfony-version: '^4.4'
           - php-version: '8.3'
             symfony-version: '^5.0'
           - php-version: '8.3'
             symfony-version: '^6.0'
+          - php-version: '8.3'
+            symfony-version: '^7.0'
           - php-version: '8.3'
             symfony-version: 'latest'
